refactor : 같은 라이프 사이클을 갖는 것으로 보이는 Menu - MenuGroup 패키지 합침
