+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -1,7 +1,6 @@
 package kitchenpos.menu.domain;
 
 import kitchenpos.menu.exception.IllegalPriceException;
-import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
